fix(agents-api): unwrap ``ActivityError`
