+++ agents-api/agents_api/activities/task_steps/base_evaluate.py
@@ -9,30 +9,12 @@ from openai import BaseModel
 # Increase the max string length to 2048000
 simpleeval.MAX_STRING_LENGTH = 2048000
 
-from simpleeval import NameNotDefined, SimpleEval
+from simpleeval import SimpleEval
 from temporalio import activity
-from thefuzz import fuzz
 
 from ...common.protocol.tasks import StepContext
 from ..utils import get_evaluator
-
-
-class EvaluateError(Exception):
-    def __init__(self, error, expression, values):
-        error_message = error.message if hasattr(error, "message") else str(error)
-        message = error_message
-
-        # Catch a possible jinja template error
-        if "unhashable" in error_message and "{{" in expression:
-            message += "\nSuggestion: It seems like you used a jinja template, did you mean to use a python expression?"
-
-        # Catch a possible misspell in a variable name
-        if isinstance(error, NameNotDefined):
-            misspelledName = error_message.split("'")[1]
-            for variableName in values:
-                if fuzz.ratio(variableName, misspelledName) >= 90.0:
-                    message += f"\nDid you mean '{variableName}' instead of '{misspelledName}'?"
-        super().__init__(message)
+from ...common.exceptions.executions import EvaluateError
 
 
 # Recursive evaluation helper function
