+++ agents-api/tests/test_base_evaluate.py
@@ -1,7 +1,8 @@
 import uuid
 from unittest.mock import patch
 
-from agents_api.activities.task_steps.base_evaluate import EvaluateError, base_evaluate
+from agents_api.activities.task_steps.base_evaluate import base_evaluate
+from agents_api.common.exceptions.executions import EvaluateError
 from agents_api.autogen.openapi_model import (
     Agent,
     TaskSpecDef,
